Added fences framework
Whole bunch of files

+++ MonoGame.Graphics.AZDO/IShaderProgramCache.cs
@@ -3,7 +3,7 @@ namespace MonoGame.Graphics.AZDO
 	public interface IShaderProgramCache
 	{
 		byte ProgramIndex { get; }
-		void SetProgram (byte programIndex);
+		void SetProgram (ushort programIndex);
 		IShaderProgram GetActiveProgram ();
 	}
