+++ MonoGame.Graphics/ShaderProgram.cs
@@ -4,7 +4,7 @@ namespace MonoGame.Graphics
 {
 	public class ShaderProgram
 	{
-		public byte ProgramIndex {get;set;}
+		public ushort ProgramIndex {get;set;}
 		public string VertexShader {get;set;}
 		public string FragmentShader { get; set; }		
 	}
